requirements.txt
# Core Dependencies
numpy>=1.24.3
pandas>=2.1.4
scikit-learn>=1.3.2
psutil>=5.9.8

# Deep Learning
tensorflow>=2.15.0
torch>=2.0.0
torchvision>=0.15.0
keras>=2.15.0

# Web Interfaces
streamlit>=1.32.0
gradio>=4.19.2

# Data Processing
nltk>=3.8.1
beautifulsoup4>=4.12.0
opencv-python>=4.8.0
Pillow>=10.2.0

# API Integration
openai>=1.0.0
google-generativeai>=0.3.2
kaggle==1.5.12

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
nest-asyncio>=1.6.0
streamlit-plotly-events>=0.0.4
streamlit-autorefresh>=0.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Added from the code block
plotly>=5.18.0
